Lab2/requirements.txt
Flask
psycopg2-binary
websockets
